+++ delphi-frontend/src/main/java/au/com/integradev/delphi/symbol/SearchMode.java
@@ -1,0 +1,6 @@
+package au.com.integradev.delphi.symbol;
+
+public enum SearchMode {
+  DEFAULT,
+  METHOD_HEADING
+}