+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Rename language key from `delph` to `delphi`.
 - Rename repository key from `delph` to `community-delphi`.
 - Rename repository from `Delphi` to `Community Delphi`.
+- Improve name resolution for declarations within types.
 
 ### Removed